SOM_CPC_requirements.txt
numpy==1.17.3
urllib3==1.25.4
pillow==8.3.2
autopep8==1.5.4
future==0.18.2
joblib==1.0.0
mne==0.21.2
pandas==1.0.1
pycodestyle==2.6.0
pyqt5-sip==4.19.18
pyqtwebengine==5.12.1
pyyaml==5.3.1
scikit-learn==0.22.2
scikit-image==0.17.2
scipy==1.5.4
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.54.1
plotly==5.4.0
tenacity==8.0.1
xlrd==2.0.1
openpyxl==3.0.9
matplotlib==3.1.3
ipykernel==5.1.3
ipython==7.9.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
h5py==2.10.0
psutil==5.9.0

